mobile dashboard card styling and html
