+++ nexuse2e-webapp/src/main/frontend/src/app/list/list.component.html
@@ -56,8 +56,8 @@
           <div *ngIf="headerElement && headerElement.linkUrlRecipe" class="card-header">
             <mat-checkbox *ngIf="isSelectable" (change)="toggleSelection($event, item)"
                           [checked]="isSelected(item)"></mat-checkbox>
-            <a [routerLink]="getUrl(item, headerElement.linkUrlRecipe)">
-              {{ getProperty(item, headerElement.fieldName) }}
+            <a [routerLink]="urlBuilderService.getUrl(item, headerElement.linkUrlRecipe)">
+              {{ urlBuilderService.getProperty(item, headerElement.fieldName) }}
             </a>
           </div>
           <table>
@@ -107,11 +107,11 @@
 
 <ng-template #cellContent let-columnConfigItem="columnConfigItem" let-item="item" let-doNotTrim="doNotTrim">
   <span class="icon-property" *ngIf="columnConfigItem.columnType == columnType().ICON">
-    <img [src]="getUrl(item, columnConfigItem.iconUrlRecipe)" alt="property_icon"/>
+    <img [src]="urlBuilderService.getUrl(item, columnConfigItem.iconUrlRecipe)" alt="property_icon"/>
   </span>
   <ng-container
     *ngIf="columnConfigItem.columnType == columnType().TEXT || columnConfigItem.columnType == columnType().TEXT_AND_MORE">
-    <span>{{ columnConfigItem.titleCase ? (getProperty(item, columnConfigItem.fieldName) | capsToTitleCase) : getProperty(item, columnConfigItem.fieldName) }}</span>
+    <span>{{ columnConfigItem.titleCase ? (urlBuilderService.getProperty(item, columnConfigItem.fieldName) | capsToTitleCase) : urlBuilderService.getProperty(item, columnConfigItem.fieldName) }}</span>
   </ng-container>
   <ng-container *ngIf="columnConfigItem.columnType == columnType().LONG_TEXT">
     <ng-container *ngIf="!doNotTrim">
@@ -122,13 +122,13 @@
       </div>
     </ng-container>
     <ng-container *ngIf="doNotTrim">
-      <span>{{ getProperty(item, columnConfigItem.fieldName) }}</span>
+      <span>{{ urlBuilderService.getProperty(item, columnConfigItem.fieldName) }}</span>
     </ng-container>
   </ng-container>
   <ng-container
     *ngIf="columnConfigItem.columnType == columnType().LINK || columnConfigItem.columnType == columnType().LINK_AND_MORE">
-    <a [routerLink]="getUrl(item, columnConfigItem.linkUrlRecipe)">
-      {{ getProperty(item, columnConfigItem.fieldName) }}
+    <a [routerLink]="urlBuilderService.getUrl(item, columnConfigItem.linkUrlRecipe)">
+      {{ urlBuilderService.getProperty(item, columnConfigItem.fieldName) }}
     </a>
   </ng-container>
   <ng-container
@@ -139,20 +139,20 @@
                 class="separator-padding">|</span>
           <span *ngIf="columnConfigItem.separator == separator().BRACKETS">(</span>
             <ng-container *ngIf="columnConfigItem.additionalLinkUrlRecipe">
-              <a [routerLink]="getUrl(item, columnConfigItem.additionalLinkUrlRecipe)"
-                 [queryParams]="getQueryParams(item, columnConfigItem.additionalLinkQueryParamsRecipe)">
-                {{ "" + columnConfigItem.additionalLinkText ? ("labels." + columnConfigItem.additionalLinkText | translate) : getProperty(item, columnConfigItem.additionalFieldName) + "" }}
+              <a [routerLink]="urlBuilderService.getUrl(item, columnConfigItem.additionalLinkUrlRecipe)"
+                 [queryParams]="urlBuilderService.getQueryParams(columnConfigItem.additionalLinkQueryParamsRecipe, item)">
+                {{ "" + columnConfigItem.additionalLinkText ? ("labels." + columnConfigItem.additionalLinkText | translate) : urlBuilderService.getProperty(item, columnConfigItem.additionalFieldName) + "" }}
               </a>
             </ng-container>
             <ng-container *ngIf="!columnConfigItem.additionalLinkUrlRecipe">
-              {{ "" + getProperty(item, columnConfigItem.additionalFieldName) + "" }}
+              {{ "" + urlBuilderService.getProperty(item, columnConfigItem.additionalFieldName) + "" }}
             </ng-container>
           <span *ngIf="columnConfigItem.separator == separator().BRACKETS">)</span>
         </span>
     </ng-container>
   </ng-container>
   <mat-icon *ngIf="columnConfigItem.showCopyIcon && showAsSimpleTable" class="copyToClipboard"
-            [cdkCopyToClipboard]="getProperty(item, columnConfigItem.fieldName)"
+            [cdkCopyToClipboard]="urlBuilderService.getProperty(item, columnConfigItem.fieldName)"
             (click)="showCopiedNotification()">content_copy
   </mat-icon>
 </ng-template>
