+++ nexuse2e-webapp/src/main/frontend/src/app/successful-messages/successful-messages.component.scss
@@ -1,3 +1,5 @@
+@import '../../styles.scss';
+
 .tab-group {
   margin-top: 10px;
