+++ nexuse2e-webapp/src/main/frontend/src/app/successful-messages/successful-messages.component.ts
@@ -1,10 +1,13 @@
 import { Component, Input, OnInit } from "@angular/core";
 import { Choreography, Partner } from "../types";
 import {
+  CARD_LINK_CONFIG,
   SUCCESS_MESS__CHOREOGRAPHY_CONFIG,
   SUCCESS_MESS__PARTNER_CONFIG,
 } from "./successful-messages.config";
 import { DataService } from "../services/data.service";
+import { ScreensizeService } from "../services/screensize.service";
+import { UrlBuilderService } from "../services/url-builder.service";
 
 @Component({
   selector: "app-successful-messages-list",
@@ -18,8 +21,14 @@ export class SuccessfulMessagesComponent implements OnInit {
   partners: Partner[] | undefined;
   partnerConfig = SUCCESS_MESS__PARTNER_CONFIG;
   loaded = false;
+  mobileCardLinkUrl = CARD_LINK_CONFIG.linkUrl;
+  mobileCardLinkQueryParams = CARD_LINK_CONFIG.linkParamsRecipe;
 
-  constructor(private dataService: DataService) {}
+  constructor(
+    private dataService: DataService,
+    public screenSizeService: ScreensizeService,
+    public urlBuilderService: UrlBuilderService
+  ) {}
 
   async ngOnInit() {
     this.partners = await this.dataService.getStatisticsPartners();
